test(agentic): split curation and persistence tools test suite and use mock helpers
